Refractor + moved sitename, contact email to environment variables
