+++ src/lib/server/rooms.model.js
@@ -1,5 +1,5 @@
 import { prisma } from "./db";
-import { timeZone } from "../../../settings";
+import { timeZone } from "../settings";
 
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc";
