+++ src/lib/settings.ts
@@ -1,5 +1,3 @@
 export { default as Logo } from "$lib/assets/img.png";
 export const timeZone = "America/New_York";
-
-export const SITE_NAME = "Template";
-export const CONTACT_EMAIL = "Email here";
+import { PUBLIC_CONTACT_EMAIL, PUBLIC_SITE_NAME } from '$env/static/public';
