+++ src/routes/(authenticated)/(admin)/api/editUserRole/+server.ts
@@ -3,14 +3,14 @@ import { json } from "@sveltejs/kit";
 import nodemailer from "nodemailer";
 import { EMAIL_PASSWORD } from "$env/static/private";
 import dayjs from "dayjs";
-import { CONTACT_EMAIL, SITE_NAME } from "../../../../../../settings.js";
+import { PUBLIC_CONTACT_EMAIL, PUBLIC_SITE_NAME } from '$env/static/public';
 
 const transporter = nodemailer.createTransport({
   host: "smtp.titan.email",
   port: 465,
   secure: true,
   auth: {
-    user: CONTACT_EMAIL,
+    user: PUBLIC_CONTACT_EMAIL,
     pass: EMAIL_PASSWORD,
   },
 });
@@ -40,10 +40,10 @@ export async function POST({ request }) {
           return json({ error: error }, { status: 400 });
         }
         const info = await transporter.sendMail({
-          from: "Support <" + CONTACT_EMAIL + ">",
+          from: "Support <" + PUBLIC_CONTACT_EMAIL + ">",
           to: "eniadebisi@gmail.com",
           subject: "Room reservation password reset",
-          html: "<div style='text-align: center;'> <h1> " + SITE_NAME + "Reset</h1> <p>Hello, " + userName + "This will be your new temporary password: '" + newTempPassword + "'</p>  <p>You will have until " + dayjs(deadline).format() + " to reset your password from now.</p></div>",
+          html: "<div style='text-align: center;'> <h1> " + PUBLIC_SITE_NAME + "Reset</h1> <p>Hello, " + userName + "This will be your new temporary password: '" + newTempPassword + "'</p>  <p>You will have until " + dayjs(deadline).format() + " to reset your password from now.</p></div>",
         });
         console.log("Message sent " + info.messageId);
       }
