+++ src/routes/(authenticated)/(admin)/api/newReservation/+server.ts
@@ -1,5 +1,4 @@
 import { checkAvailability, reserveRoom } from "$lib/server/rooms.model";
-import { timeZone } from "../../../../../../settings";
 import { json } from "@sveltejs/kit";
 import dayjs from "dayjs";
