+++ src/routes/(authenticated)/(admin)/api/reservationData/+server.ts
@@ -1,5 +1,5 @@
 import { getReservations } from "$lib/server/rooms.model";
-import { timeZone } from "../../../../../../settings";
+import { timeZone } from "$lib/settings";
 import { json } from "@sveltejs/kit";
 import dayjs from "dayjs";
 
@@ -10,8 +10,8 @@ export async function POST({ request }) {
     return json({ error: true }, { status: 400 });
   }
 
-  let start = dayjs(new Date(nDate)).tz(timeZone).startOf("day").utc()
-  let end = dayjs(new Date(nDate)).tz(timeZone).endOf("day").utc()
+  let start = dayjs(new Date(nDate)).tz(timeZone).startOf("day").utc();
+  let end = dayjs(new Date(nDate)).tz(timeZone).endOf("day").utc();
   const { reservations, error: reservError } = await getReservations(start, end);
   if (reservError || !reservations) throw new Error();
