+++ src/routes/(authenticated)/profile/+page.server.ts
@@ -1,4 +1,4 @@
-r// import { checkSignIn } from "$lib/server/user.model";
+// import { checkSignIn } from "$lib/server/user.model";
 // import type { PageServerLoad } from "./$types";
 // import type { Actions } from "./$types";
 // import { error, fail, redirect } from "@sveltejs/kit";
