+++ src/routes/register/+page.server.ts
@@ -3,8 +3,8 @@ import { signUp } from "$lib/server/user.model";
 import type { Actions } from "./$types";
 import { fail, redirect } from "@sveltejs/kit";
 
-export async function load({ locals }) {
-  if (!locals.user || locals.user && !hasPermission(locals.user, "UserManagement", "create")) redirect(302, "/reservations");
+export async function load({ locals }) {  
+  if (locals.user && !hasPermission(locals.user, "UserManagement", "create")) redirect(302, "/reservations");
 
   return { user: locals.user };
 }
