+++ tsconfig.json
@@ -119,5 +119,5 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "include": ["src", "settings.ts"]
+  "include": ["src", "src/lib/settings.tss.ts"]
 }
